// src/interfaces/dashboardInterface.ts
export interface TopSectionData {
	title: string;
	pageTitle: string;
	rightClickBtn: () => void;
}

export interface WidgetObject {
    id: number;
    cardColor: string;
    label: string;
    badge: string;
    badgeClass: string;
    percentage: string;
    counter: number;
    link: string;
    bgcolor: string;
    icon: string;
    decimals: number;
    prefix: string;
    suffix: string;
    linkTo: string;
  }